Splitting components, better route navigation
This also allows us to generate a password reset link to quickly reset
your password without entering the code and email manually.
